+++ lib/scorched/response.rb
@@ -5,26 +5,32 @@ module Scorched
     def merge!(response)
       return self if response == self
       if Rack::Response === response
-        response.finish
-        self.status = response.status
-        self.header.merge!(response.header)
-        self.body = []
-        response.each { |v| self.body << v }
-      else
-        self.status, @header, self.body = response
+        response = [response.status, response.header, response]
       end
+      self.status, self.body = response[0], response[2]
+      self.header.merge!(response[1])
+      self
     end
     
     # Automatically wraps the assigned value in an array if it doesn't respond to ``each``.
     # Also filters out non-true values and empty strings.
     def body=(value)
-      value = []  if !value || value == ''
+      value = [] if !value || value == ''
       super(value.respond_to?(:each) ? value : [value.to_s])
     end
     
+    # Override finish to avoid using BodyProxy
     def finish(*args, &block)
       self['Content-Type'] ||= 'text/html;charset=utf-8'
-      super
+      @block = block if block
+      if [204, 205, 304].include?(status.to_i)
+        header.delete "Content-Type"
+        header.delete "Content-Length"
+        close
+        [status.to_i, header, []]
+      else
+        [status.to_i, header, body]
+      end
     end
     
     alias :to_a :finish